server/requirement.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
jinja2
cryptography
